+++ scripts/Test-Template.ps1
@@ -2,7 +2,7 @@ function Test-Template($templateName, $templateArgs, $templateNupkg, $isSPA) {
     $tmpDir = "$PSScriptRoot/$templateName"
     Remove-Item -Path $tmpDir -Recurse -ErrorAction Ignore
 
-    #& "$PSScriptRoot/../build.cmd" /t:Package
+    & "$PSScriptRoot/../build.cmd" /t:Package
     Run-DotnetNew "--install", "$PSScriptRoot/../artifacts/build/$templateNupkg"
 
     New-Item -ErrorAction Ignore -Path $tmpDir -ItemType Directory
