+++ test/Templates.Test/PuppeteerTests/__tests__/reactreduxTest.ts
@@ -1,7 +1,7 @@
 import { Page, Browser, launch } from 'puppeteer';
-import { bindConsole, clickByText, validateMessages } from '../../testFuncs/testFuncs';
+import { bindConsole, clickByText, validateMessages } from '../testFuncs';
 
-const serverPath = `https://localhost:6021`;
+const serverPath = `https://localhost:6031`;
 
 jest.setTimeout(30000);
 
@@ -26,7 +26,7 @@ afterAll(async () => {
     }
 });
 
-describe('react pages are ok', () => {
+describe('reactredux pages are ok', () => {
     it('all pages work', async () => {
         await page.goto(serverPath);
         await page.waitFor('ul');
@@ -54,7 +54,9 @@ describe('react pages are ok', () => {
 
         heading = await page.$eval('h1', heading => heading.textContent);
         expect(heading).toBe('Weather forecast');
-        await page.waitFor(200);
+        await page.waitForSelector("td");
+        await page.waitForFunction('document.querySelector("td").innerText.length > 0');
+
         let trs = await page.$x('//tbody//tr');
 
         expect(trs.length).toBe(5);
