+++ test/Templates.Test/PuppeteerTests/package.json
@@ -12,11 +12,16 @@
     "puppeteer": "^1.8.0",
     "@types/puppeteer": "^1.9.1",
     "typescript-collections": "^1.3.2",
-    "os": "^0.1.1",
-    "puppeteer-tests": "file:../testFuncs"
+    "os": "^0.1.1"
   },
   "scripts": {
-    "test": "jest"
+    "test-angular": "jest __test__/angularTest.ts",
+    "test-mvc": "jest __test__/mvcTest.ts",
+    "test-razor": "jest __test__/razorTest.ts",
+    "test-react": "jest __test__/reactTest.ts",
+    "test-reactredux": "jest __test__/reactreduxTest.ts",
+    "test-webapi": "jest __tests__/webapiTest.ts",
+    "test-web": "jest __tests__/webTest.ts"
   },
   "jest": {
     "moduleFileExtensions": [
