--- test/Templates.Test/PuppeteerTests/reactredux/package.json
@@ -1,40 +1,0 @@
-{
-  "private": true,
-  "version": "0.0.0",
-  "dependencies": {
-    "typescript": "*"
-  },
-  "devDependencies": {
-    "@types/jest": "^23.3.9",
-    "@types/node": "^10.12.2",
-    "jest": "^23.6.0",
-    "ts-jest": "^23.10.4",
-    "puppeteer": "^1.8.0",
-    "@types/puppeteer": "^1.9.1",
-    "typescript-collections": "^1.3.2",
-    "os": "^0.1.1",
-    "puppeteer-tests": "file:../testFuncs"
-  },
-  "scripts": {
-    "test": "jest"
-  },
-  "jest": {
-    "moduleFileExtensions": [
-      "json",
-      "ts",
-      "tsx",
-      "js"
-    ],
-    "transform": {
-      "^.+\\.(ts|tsx)$": "ts-jest"
-    },
-    "globals": {
-      "ts-jest": {
-        "tsConfig": "tsconfig.json"
-      }
-    },
-    "testMatch": [
-      "**/__tests__/*.+(ts|tsx|js)"
-    ]
-  }
-}
