--- test/Templates.Test/PuppeteerTests/reactredux/tsconfig.json
@@ -1,7 +1,0 @@
-{
-    "compilerOptions": {
-        "target": "ES2015",
-        "esModuleInterop": true
-    },
-    "preset": "ts-test"
-}
