--- test/Templates.Test/PuppeteerTests/testFuncs/package.json
@@ -1,5 +1,0 @@
-{
-  "devDependencies": {
-    "puppeteer": "^1.8.0"
-  }
-}
