--- test/Templates.Test/PuppeteerTests/web/tsconfig.json
@@ -1,7 +1,0 @@
-{
-    "compilerOptions": {
-        "target": "ES2015",
-        "esModuleInterop": true
-    },
-    "preset": "ts-test"
-}
